requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
faiss-cpu>=1.7.0
Pillow>=8.0.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-learn>=1.0.0
snntorch>=0.5.0  # Optional, for advanced SNN features
